backend/requirements.txt
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
Flask==3.0.0
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
PyMySQL==1.1.0
SQLAlchemy==2.0.24
typing_extensions==4.9.0
Werkzeug==3.0.1
flask_login
requests_oauthlib
python-dotenv
pandas
